Add support for sorting in Tree mode
Sort denotes to children.

Collap button renamed to SortBy.
In order to not create any more buttons.
Keybinding F6 changed to SortBy as well. The rest bindings
are not changed so you can collapse elementes by '+', '-'.

The commit wasn't tested by tests since I am not sure if they exists.
But if they are they may be deserve be updated.